pyproject.toml
[project]
name = "packfiles"
version = "0.1.0"
description = "packfiles allows merging multiple files into a single Markdown file."
authors = [{ name = "ecmaxp", email = "ecmaxp@ecmaxp.kr" }]
requires-python = ">=3.10"
dependencies = ["ignorelib>=0.3.0", "python-magic>=0.4.27"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
    "Topic :: System :: Archiving",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]

[dependency-groups]
dev = ["reloader-py>=0.16.3"]

[project.urls]
Repository = "https://github.com/EcmaXp/packfiles"

[project.scripts]
"packfiles" = "packfiles:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packfiles.py"]

[tool.hatch.build.targets.sdist]
packages = ["packfiles.py"]

[tool.pyright]
venvPath = "."
venv = ".venv"
